test(Breeze.ChildServer): supervise processes and speed up test runs
Run large test groups and example loading concurrently, and start ChildServer,
Server, and InputRouter processes under ExUnit supervision. Coordinate
timing-sensitive InputRouter tests explicitly and clean up profiler scopes,
reloaders, watchers, and ETS ownership so repeated suites remain fast and
memory-stable.
