+++ lib/breeze/debug_profiler.ex
@@ -5,11 +5,13 @@ defmodule Breeze.DebugProfiler do
   @span_event [:breeze, :render]
   @handler_id "breeze-debug-profiler"
   @table :breeze_debug_profile
+  @tracked_table_owner_key {__MODULE__, :tracked_table_owner}
 
   def reset(scope) do
     if enabled?() do
       ensure_handler()
       delete_scope(scope)
+      activate_scope(scope)
     end
 
     :ok
@@ -30,6 +32,21 @@ defmodule Breeze.DebugProfiler do
     else
       []
     end
+  end
+
+  def discard(scope) do
+    delete_scope(scope)
+    :ok
+  end
+
+  def discard_owner(owner) when is_pid(owner) do
+    with_table(fn ->
+      @table
+      |> :ets.match_object({{:active, :_}, owner})
+      |> Enum.each(fn {{:active, scope}, ^owner} -> delete_scope_entries(scope) end)
+    end)
+
+    :ok
   end
 
   def enabled? do
@@ -63,23 +80,50 @@ defmodule Breeze.DebugProfiler do
   end
 
   defp delete_scope(scope) do
+    with_table(fn -> delete_scope_entries(scope) end)
+  end
+
+  defp delete_scope_entries(scope) do
+    :ets.select_delete(
+      @table,
+      [
+        {
+          {{scope, :"$1", :"$2"}, :_},
+          [],
+          [true]
+        }
+      ]
+    )
+
+    :ets.delete(@table, {:active, scope})
+  end
+
+  defp activate_scope(scope) do
+    track_scope_owner()
+
     with_table(fn ->
-      :ets.select_delete(
-        @table,
-        [
-          {
-            {{scope, :"$1", :"$2"}, :_},
-            [],
-            [true]
-          }
-        ]
-      )
+      :ets.insert(@table, {{:active, scope}, self()})
     end)
+  end
+
+  defp track_scope_owner do
+    pid =
+      case Process.whereis(__MODULE__.TableOwner) do
+        nil -> ensure_table_owner()
+        pid -> pid
+      end
+
+    unless Process.get(@tracked_table_owner_key) == pid do
+      __MODULE__.TableOwner.track_owner(pid, self())
+      Process.put(@tracked_table_owner_key, pid)
+    end
   end
 
   defp insert_metric(scope, label, metric, value) do
     with_table(fn ->
-      :ets.insert(@table, {{scope, label, metric}, value})
+      if :ets.member(@table, {:active, scope}) do
+        :ets.insert(@table, {{scope, label, metric}, value})
+      end
     end)
 
     :ok
@@ -121,7 +165,8 @@ defmodule Breeze.DebugProfiler do
           pid
       end
 
-    __MODULE__.TableOwner.ensure_table(pid)
+    :ok = __MODULE__.TableOwner.ensure_table(pid)
+    pid
   end
 
   def handle_event(
@@ -162,16 +207,48 @@ defmodule Breeze.DebugProfiler.TableOwner do
     GenServer.call(pid, :ensure_table)
   end
 
+  def track_owner(pid, owner) do
+    GenServer.cast(pid, {:track_owner, owner})
+  end
+
   @impl true
-  def init(state) do
+  def init(_state) do
     ensure_table!()
-    {:ok, state}
+    {:ok, %{owners: %{}, owners_by_ref: %{}}}
   end
 
   @impl true
   def handle_call(:ensure_table, _from, state) do
     ensure_table!()
     {:reply, :ok, state}
+  end
+
+  @impl true
+  def handle_cast({:track_owner, owner}, state) do
+    if Map.has_key?(state.owners, owner) do
+      {:noreply, state}
+    else
+      ref = Process.monitor(owner)
+
+      {:noreply,
+       %{
+         state
+         | owners: Map.put(state.owners, owner, ref),
+           owners_by_ref: Map.put(state.owners_by_ref, ref, owner)
+       }}
+    end
+  end
+
+  @impl true
+  def handle_info({:DOWN, ref, :process, owner, _reason}, state) do
+    Breeze.DebugProfiler.discard_owner(owner)
+
+    {:noreply,
+     %{
+       state
+       | owners: Map.delete(state.owners, owner),
+         owners_by_ref: Map.delete(state.owners_by_ref, ref)
+     }}
   end
 
   defp ensure_table! do
